Extract the pure recognition core to _lift.py
recognized_tile(op, output) — softmax fusion, the free-axis peel, the cell
lift / nodification — is the schedule fork's root op on the fused arm, and
therefore the carrier of a kernel's structural identity. Factoring it out of
the 010_recognize pass shell (which keeps the placement fork and cut
realization) lets the strict golden decode derive a RECORD's identity through
the exact functions the live compile recognizes with: one shared code path, so
record-side and fork-side identity cannot drift. Byte-identical kernels by the
pinned digest gate; the boundary-rule tests import the family from its new
home instead of importlib-loading the digit module.
