+++ emmy/compiler/pipeline/passes/lowering/kernel/010_materialize.py
@@ -6,7 +6,7 @@ cooperative / ILP reduce, or a pointwise / scalar cell) lowers through that sing
 reads the node kind + role + reduce plan off ``tile.op`` and picks the tier:
 
 - **Tiled ``CONTRACTION``** (warp / register tile) — the high-level contraction was
-  built recognize-side (``010_recognize._nodify_contraction``); ``factorize`` synthesizes its bare
+  built recognize-side (``_lift._nodify_contraction``); ``factorize`` synthesizes its bare
   grid-``Write`` (needs ``root.output``, so it can't ride the node) and expands it (mma → the
   ``RegFragment`` / ``LdmatrixLoad`` / ``MmaSyncPtx`` / ``RegStore`` fragment soup; scalar → the
   per-thread register cell tile) through the shared four-level tiling layer (in ``_factor``).
